FrontEnd Java Doc V1 and Clean Up

+++ src/commands/RandomCommand.java
@@ -6,13 +6,12 @@ import controller.MasterController;
 public class RandomCommand extends OneInputCommand {
 
     @Override
-    public double executeCommand(Model model) {
-    	try{
-    		return (double)MasterController.myRandom.nextInt((int)myParameters[0].executeCommand(model));
-    	}
-    	catch(IllegalArgumentException e){
-    		System.out.println("Exception caught in RandomCommand: ZERO OR LESS");
-    		return 0;
-    	}
+    public double executeCommand (Model model) {
+        try {
+            return MasterController.myRandom.nextInt((int) myParameters[0].executeCommand(model));
+        } catch (IllegalArgumentException e) {
+            System.out.println("Exception caught in RandomCommand: ZERO OR LESS");
+            return 0;
+        }
     }
 }
